% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/highlight.R
\name{as_dict}
\alias{as_dict}
\title{Construct dictionary}
\usage{
as_dict(x, feature = "feature", bg_colour = "bg_colour",
  txt_colour = "txt_colour", bold = "bold", italic = "italic",
  strike_out = "strike_out")
}
\arguments{
\item{x}{A data.frame with columns of switches to determine text
highlighting. Not all columns need to be present.}

\item{feature}{Name of column that contains the words of terms to highlight.}

\item{bg_colour}{Name of column that controls the background colour for a
feature}

\item{txt_colour}{Name of column that controls the text colour for a feature.}

\item{bold}{Name of column that controls if a feature is printed with bold
font.}

\item{italic}{Name of column that controls if a feature is printed with
italic font.}

\item{strike_out}{Name of column that controls if a feature is printed
struck out.}
}
\description{
Construct a dictionary which controls the highlighting of text.
}
\details{
The columns "bold", "italic" and "strike_out" need to contain
logical values (i.e. TRUE/FALSE)
}
\examples{
df <- data.frame(
  feature = c("good", "bad"),
  bg_colour = c("green", "red"),
  txt_colour = c("grey", "blue"),
  bold = TRUE,
  italic = TRUE,
  strike_out = TRUE,
  stringsAsFactors = FALSE
)
mydict <- as_dict(df)
}
