% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/highlight.R
\name{highlight}
\alias{highlight}
\title{Text highlighter}
\usage{
highlight(text, dict, id = NULL, case_insensitive = TRUE,
  output = c("html", "tex"))
}
\arguments{
\item{text}{A text to be printed with highlighted words.}

\item{dict}{A data.frame with columns "word" and "colour".}

\item{id}{A vector of ids which are displayed as headlines.}

\item{case_insensitive}{Should search for feature be case insensitive?}

\item{output}{Choose output format.}
}
\description{
Text highlighter
}
\examples{
text <- c("This is a good test with some bad words", "bad is red and good is green")
df <- data.frame(
  feature = c("good", "bad"),
  bg_colour = c("green", "red"),
  txt_colour = c("grey", "blue"),
  bold = TRUE,
  italic = TRUE,
  strike_out = TRUE,
  stringsAsFactors = FALSE
)
mydict <- as_dict(df)
highlight(text, mydict)

}
