% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{notepad_example}
\alias{notepad_example}
\title{Comment Example Dataset}
\format{
\subsection{\code{notepad_example}}{

A data frame with 126 rows and 2 columns:
\describe{
\item{ID}{Participant Identifier, as well as source document identifier}
\item{Text}{Participant notes, as well as source transcript}

}
}
}
\source{
Jury Perception Study (see Rogers (2024) \url{https://digitalcommons.unl.edu/dissertations/AAI31240449/})
}
\usage{
notepad_example
}
\description{
Participant comments for the initial description used in the jury perception
study
}
\keyword{datasets}
