% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/token_comments.R
\name{token_comments}
\alias{token_comments}
\title{Tokenize comments}
\usage{
token_comments(comment_document)
}
\arguments{
\item{comment_document}{document containing comments}
}
\value{
tokenized comments
}
\description{
Tokenize comments
}
