% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/token_transcript.R
\name{token_transcript}
\alias{token_transcript}
\title{Tokenize Transcript}
\usage{
token_transcript(transcript_file)
}
\arguments{
\item{transcript_file}{data frame of the transcript}
}
\value{
a tokenized object
}
\description{
Tokenize Transcript
}
\examples{
transcript_example_rename <- dplyr::rename(transcript_example, text=Text)
toks_transcript <- token_transcript(transcript_example_rename)
}
