% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tokenize_derivate.R
\name{tokenize_derivative}
\alias{tokenize_derivative}
\title{Tokenize comments}
\usage{
tokenize_derivative(tbl, source_row, text_column)
}
\arguments{
\item{tbl}{data frame containing documents, where each
row represents a document}

\item{source_row}{row containing text to be treated as source}

\item{text_column}{string indicating the name of the column containing derivative text}
}
\value{
tokenized comments
}
\description{
This function tokenizes comments that are to be used in \code{\link[=collocation_frequency]{collocation_frequency()}}
}
\examples{
# Tokenize the derivative document
src_row <- which(notepad_example$ID=="source")
toks_comment <- tokenize_derivative(notepad_example, source_row=src_row, text_column="Text")
}
