% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transcript_cleaning.R
\name{transcript_cleaning}
\alias{transcript_cleaning}
\title{Transcript Cleaning Function}
\usage{
transcript_cleaning(transcript)
}
\arguments{
\item{transcript}{transcript document to be read (assumed to be a single page)}
}
\value{
the transcript broken down by words
}
\description{
Transcript Cleaning Function
}
