% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{transcript_example}
\alias{transcript_example}
\title{Transcript Example}
\format{
\subsection{\code{validation_dataset}}{

A data frame with 561 rows and 4 columns:
\describe{
\item{clean_prints}{Identifier}
\item{page_count}{page number}
\item{notes}{The uncleaned notes}
\item{corrected_notes}{hand cleaned notes to demonstrate correct format}

}
}
}
\source{
Jury Perception Study
}
\usage{
transcript_example
}
\description{
Text corresponding to participant comments
}
\keyword{datasets}
