% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transcript_frequency.R
\name{transcript_frequency}
\alias{transcript_frequency}
\title{Mapping Collocation Frequency to Transcript Document}
\usage{
transcript_frequency(transcript, collocate_object)
}
\arguments{
\item{transcript}{transcript document}

\item{collocate_object}{collocation object (returned
from \code{\link[=collocate_comments_fuzzy]{collocate_comments_fuzzy()}} or \code{\link[=collocate_comments]{collocate_comments()}})}
}
\value{
a dataframe of the transcript document with collocation values by word
}
\description{
This function connects the collocation frequency calculated in
\code{\link[=collocate_comments_fuzzy]{collocate_comments_fuzzy()}} to the base transcript.
}
\examples{
# Rename relevant column to page_notes in the derivative document
comment_example_rename <- dplyr::rename(comment_example, page_notes=Notes)
# Tokenize the derivative document
toks_comment <- tokenize_derivative(comment_example, text_column="Notes")
# Tokenize source document
toks_source <- tokenize_source(as.character(transcript_example))
# Compute collocation frequencies
collocation_object <- collocate_comments(toks_source, toks_comment)
# Merge frequencies with source document to provide averages by word and correct formatting
merged_frequency <- transcript_frequency(as.character(transcript_example), collocation_object)
}
