% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{wiki_pages}
\alias{wiki_pages}
\title{Wikipedia Edit History for "Highlighter"}
\format{
\subsection{\code{wiki_pages}}{

A data frame with 50 rows and 1 column:
\describe{
\item{page_notes}{text of the Wikipedia page for Highlighter}

}
}
}
\source{
Wikipedia: https://en.wikipedia.org/w/index.php?title=Highlighter&action=history
}
\usage{
wiki_pages
}
\description{
Text corresponding to versions of the Wikipedia article for Highlighter
}
\keyword{datasets}
