% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collocate_comments.R
\name{collocate_comments}
\alias{collocate_comments}
\title{Collocation of Comments}
\usage{
collocate_comments(transcript_token, note_token)
}
\arguments{
\item{transcript_token}{transcript token to act as baseline for notes, resulting
from \code{\link[=token_transcript]{token_transcript()}}}

\item{note_token}{tokenized document of notes, resulting from \code{\link[=token_comments]{token_comments()}}}
}
\value{
data frame of transcript and corresponding note frequency
}
\description{
This function provides the frequency of collocations in comments that
correspond to the provided transcript.
}
