% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collocate_comments_fuzzy.R
\name{collocate_comments_fuzzy}
\alias{collocate_comments_fuzzy}
\title{Collocate Comments Fuzzy}
\usage{
collocate_comments_fuzzy(transcript_token, note_token)
}
\arguments{
\item{transcript_token}{transcript token to act as baseline for notes}

\item{note_token}{tokenized document of notes}
}
\value{
data frame of transcript and corresponding note frequency
}
\description{
Collocate Comments Fuzzy
}
