% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collocation_plot.R
\name{collocation_plot}
\alias{collocation_plot}
\title{Map collocation to ggplot object}
\usage{
collocation_plot(
  frequency_doc,
  colors = c("#f251fc", "#f8ff1b"),
  values = "Freq",
  order = "word_num",
  text = "words"
)
}
\arguments{
\item{frequency_doc}{document of frequencies (returned from
\code{\link[=collocation_frequency]{collocation_frequency()}})}

\item{colors}{list for color specification for the gradient. Default is c("#f251fc","#f8ff1b")}

\item{values}{column name of values to use in gradient calculation. Default is "Freq",
corresponding to document returned from \code{\link[=collocation_frequency]{collocation_frequency()}}}

\item{order}{column name corresponding to the the word order of the text. Default
is "word_num", corresponding to the document returned from \code{\link[=collocation_frequency]{collocation_frequency()}}}

\item{text}{column name corresponding to text to map the gradient to. Default is "words",
corresponding to the document returned from \code{\link[=collocation_frequency]{collocation_frequency()}}}
}
\value{
list of plot, plot object, and frequency
}
\description{
This assigns colors based on frequency to the words in the transcript.
}
\examples{
# Identify Source Row
src_row <- which(notepad_example$ID=="source")
merged_frequency <- collocation_frequency(notepad_example, src_row, "Text")
# Create a plot object to assign colors based on frequency
freq_plot <- collocation_plot(merged_frequency)
}
