% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collocation_plot.R
\name{collocation_plot}
\alias{collocation_plot}
\title{Map collocation to ggplot object}
\usage{
collocation_plot(frequency_doc, n_scenario = 1)
}
\arguments{
\item{frequency_doc}{document of frequencies (returned from transcript_frequency)}

\item{n_scenario}{number of scenarios for which this transcript appeared}
}
\value{
list of plot, plot object, and frequency
}
\description{
Map collocation to ggplot object
}
