% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{comment_example}
\alias{comment_example}
\title{Comment Example Dataset}
\format{
\subsection{\code{comment_example}}{

A data frame with 125 rows and 2 columns:
\describe{
\item{ID}{Participant Identifier}
\item{Notes}{Participant notes}

}
}
}
\source{
Jury Perception Study (see Rogers (2024) \url{https://digitalcommons.unl.edu/dissertations/AAI31240449/})
}
\usage{
comment_example
}
\description{
Participant comments for the initial description used in the jury perception
study
}
\keyword{datasets}
