% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/highlighted_text.R
\name{highlighted_text}
\alias{highlighted_text}
\title{Create Highlighted Testimony}
\usage{
highlighted_text(plot_object, descript)
}
\arguments{
\item{plot_object}{plot object resulting from collocation_plot}

\item{descript}{collocation frequency document resulting from transcript_frequency}
}
\value{
css code for highlighted text
}
\description{
Create Highlighted Testimony
}
