% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/token_comments.R
\name{token_comments}
\alias{token_comments}
\title{Tokenize comments}
\usage{
token_comments(comment_document)
}
\arguments{
\item{comment_document}{document containing notes by individual, where the
column containing the notes is named page_notes}
}
\value{
tokenized comments
}
\description{
This function tokenizes comments that are to be used in \code{\link[=collocate_comments_fuzzy]{collocate_comments_fuzzy()}}
or \code{\link[=collocate_comments]{collocate_comments()}}
}
\examples{
comment_example_rename <- dplyr::rename(comment_example, page_notes=Notes)
toks_comment <- token_comments(comment_example_rename)
}
