% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/token_transcript.R
\name{token_transcript}
\alias{token_transcript}
\title{Tokenize Transcript}
\usage{
token_transcript(transcript_file)
}
\arguments{
\item{filelocation}{data frame of the transcript}
}
\value{
a tokenized object
}
\description{
Tokenize Transcript
}
