% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/token_transcript.R
\name{token_transcript}
\alias{token_transcript}
\title{Tokenize Transcript}
\usage{
token_transcript(transcript_file)
}
\arguments{
\item{transcript_file}{data frame of the transcript, where the transcript text
is in a column named text.}
}
\value{
a tokenized object
}
\description{
This function tokenizes a transcript document that is to be used in
\code{\link[=collocate_comments_fuzzy]{collocate_comments_fuzzy()}} or \code{\link[=collocate_comments]{collocate_comments()}}
}
\examples{
transcript_example_rename <- dplyr::rename(transcript_example, text=Text)
toks_transcript <- token_transcript(transcript_example_rename)
}
