% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tokenize_derivate.R
\name{tokenize_derivative}
\alias{tokenize_derivative}
\title{Tokenize comments}
\usage{
tokenize_derivative(derivative_document, text_column)
}
\arguments{
\item{derivative_document}{data frame containing derivative documents, where each
row represents a document}

\item{text_column}{string indicating the name of the column containing derivative text}
}
\value{
tokenized comments
}
\description{
This function tokenizes comments that are to be used in \code{\link[=collocation_frequency]{collocation_frequency()}}
}
\examples{
# Tokenize the derivative document
toks_comment <- tokenize_derivative(comment_example, text_column="Notes")
}
