% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tokenize_derivate.R
\name{tokenize_derivative}
\alias{tokenize_derivative}
\title{Tokenize comments}
\usage{
tokenize_derivative(comment_document)
}
\arguments{
\item{comment_document}{document containing notes by individual, where the
column containing the notes is named page_notes}
}
\value{
tokenized comments
}
\description{
This function tokenizes comments that are to be used in \code{\link[=collocate_comments_fuzzy]{collocate_comments_fuzzy()}}
or \code{\link[=collocate_comments]{collocate_comments()}}
}
\examples{
# Rename relevant column to page_notes in the derivative document
comment_example_rename <- dplyr::rename(comment_example, page_notes=Notes)
# Tokenize the derivative document
toks_comment <- tokenize_derivative(comment_example_rename)
}
