% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tokenize_source.R
\name{tokenize_source}
\alias{tokenize_source}
\title{Tokenize Source Document}
\usage{
tokenize_source(tbl, source_row, text_column)
}
\arguments{
\item{tbl}{data frame containing documents, where each
row represents a document}

\item{source_row}{row containing text to be treated as source}

\item{text_column}{string indicating the name of the column containing derivative text}
}
\value{
a tokenized object
}
\description{
This function tokenizes a source document that is to be used in
\code{\link[=collocation_frequency]{collocation_frequency()}}
}
\examples{
# Tokenize source document
src_row <- which(notepad_example$ID=="source")
toks_source <- tokenize_source(notepad_example, source_row=src_row, text_column = "Text")
}
