% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tokenize_source.R
\name{tokenize_source}
\alias{tokenize_source}
\title{Tokenize Source Document}
\usage{
tokenize_source(transcript_file)
}
\arguments{
\item{transcript_file}{data frame of the source document, where the source document text
is in a column named text.}
}
\value{
a tokenized object
}
\description{
This function tokenizes a source document that is to be used in
\code{\link[=collocate_comments_fuzzy]{collocate_comments_fuzzy()}} or \code{\link[=collocate_comments]{collocate_comments()}}
}
\examples{
# Rename relevant column in the source document to text
source_example_rename <- dplyr::rename(transcript_example, text=Text)
# Tokenize source document
toks_source <- tokenize_source(source_example_rename)
}
