% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{transcript_example}
\alias{transcript_example}
\title{Transcript Example}
\format{
\subsection{\code{transcript_example}}{

A data frame with 1 row and 1 column:
\describe{
\item{Text}{Transcript text corresponding to the jury perception study}

}
}
}
\source{
Jury Perception Study
transcript_example <- gsub("\if{html}{\out{</br>}}","\if{html}{\out{<br>}}", transcript_example)
transcript_example <- data.frame(Text = transcript_example)
}
\usage{
transcript_example
}
\description{
Text corresponding to participant comments
}
\keyword{datasets}
