% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{transcript_example}
\alias{transcript_example}
\title{Transcript Example}
\format{
\subsection{\code{transcript_example}}{

A data frame with 1 row and 1 column:
\describe{
\item{Text}{Transcript text corresponding to the jury perception study}

}
}
}
\source{
Jury Perception Study (see Rogers (2024) \url{https://digitalcommons.unl.edu/dissertations/AAI31240449/} and Garrett et. al. (2020) \url{doi:10.1037/lhb0000423})
}
\usage{
transcript_example
}
\description{
Text corresponding to participant comments
}
\keyword{datasets}
